+++ test/integration/app/app/models/comment.rb
@@ -1,0 +1,5 @@
+class Comment < ApplicationRecord
+  belongs_to :task
+
+  scoped_to_account through: :task
+end