+++ test/integration/app/app/models/task.rb
@@ -1,4 +1,7 @@
 class Task < ApplicationRecord
   belongs_to :project
-  include AccountScoped
+  belongs_to :tag, optional: true
+  has_many :comments, dependent: :destroy
+
+  scoped_to_account through: :project
 end